test(order): refactor test folder
