+++ app/src/androidTest/java/org/flyve/inventory/agent/AboutModelTest.java
@@ -30,7 +30,7 @@ import static org.mockito.Mockito.mock;
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
  * ------------------------------------------------------------------------------
- * @author    rafaelhernandez
+ * @author    @rafaelje
  * @date      23/6/18
  * @copyright Copyright © 2018 Teclib. All rights reserved.
  * @license   GPLv3 https://www.gnu.org/licenses/gpl-3.0.html
