+++ app/src/androidTest/java/org/flyve/inventory/agent/ReportModelTest.java
@@ -1,15 +1,19 @@
-package org.flyve.inventory.agent.core.splash;
+package org.flyve.inventory.agent;
 
 import android.app.Activity;
 import android.support.test.rule.ActivityTestRule;
 import android.support.test.runner.AndroidJUnit4;
+import android.support.v7.widget.RecyclerView;
 
+import org.flyve.inventory.agent.core.report.Report;
+import org.flyve.inventory.agent.core.report.ReportModel;
 import org.flyve.inventory.agent.ui.ActivityMain;
-import org.flyve.inventory.agent.ui.ActivitySplash;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
+import static android.support.test.InstrumentationRegistry.getInstrumentation;
+import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.mock;
 
 /*
@@ -40,25 +44,33 @@ import static org.mockito.Mockito.mock;
  */
 
 @RunWith(AndroidJUnit4.class)
-public class SplashModelTest {
+public class ReportModelTest {
 
     @Rule
-    public ActivityTestRule<ActivitySplash> rule  = new ActivityTestRule<>(ActivitySplash.class);
-
+    public ActivityTestRule<ActivityMain> rule  = new ActivityTestRule<>(ActivityMain.class);
 
     @Test
-    public void setupStorage() {
+    public void generateReport() {
         Activity activity = rule.getActivity();
-        Splash.Presenter presenter = mock(Splash.Presenter.class);
-        SplashModel splashModel = new SplashModel(presenter);
-        splashModel.setupStorage(activity);
+        Report.Presenter presenter = mock(Report.Presenter.class);
+        ReportModel reportModel = new ReportModel(presenter);
+        RecyclerView lst = mock(RecyclerView.class);
+        reportModel.generateReport(activity, lst);
     }
 
     @Test
-    public void nextActivityWithDelay() {
-        Activity activity = rule.getActivity();
-        Splash.Presenter presenter = mock(Splash.Presenter.class);
-        SplashModel splashModel = new SplashModel(presenter);
-        splashModel.nextActivityWithDelay(1, activity, ActivityMain.class);
+    public void showDialogShare() {
+        final Activity activity = rule.getActivity();
+        Report.Presenter presenter = mock(Report.Presenter.class);
+        final ReportModel reportModel = new ReportModel(presenter);
+
+        getInstrumentation().runOnMainSync(new Runnable() {
+            @Override
+            public void run() {
+                reportModel.showDialogShare(activity);
+            }
+        });
+
+        assertTrue(true);
     }
 }
